# test(benchmark): add queue performance check