--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -58,6 +58,7 @@ jobs:
       - name: Run Tests
         run: |
           go test -v -covermode=atomic -coverprofile=coverage.out
+          go test -v -run=^$ -benchmem -bench .
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3