--- a/go.mod
+++ b/go.mod
@@ -3,13 +3,14 @@ module github.com/golang-queue/queue
 go 1.18
 
 require (
+	github.com/goccy/go-json v0.9.7
 	github.com/golang/mock v1.6.0
 	github.com/stretchr/testify v1.7.1
 	go.uber.org/goleak v1.1.12
 )
 
 require (
-	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )