--- a/queue_test.go
+++ b/queue_test.go
@@ -1,6 +1,7 @@
 package queue
 
 import (
+	"context"
 	"testing"
 	"time"
 
@@ -139,4 +140,26 @@ func TestCloseQueueAfterShutdown(t *testing.T) {
 	})
 	assert.Error(t, err)
 	assert.Equal(t, ErrQueueShutdown, err)
+}
+
+func BenchmarkQueueTask(b *testing.B) {
+	q := NewPool(5)
+	q.Release()
+	for n := 0; n < b.N; n++ {
+		_ = q.QueueTask(func(context.Context) error {
+			time.Sleep(10 * time.Millisecond)
+			return nil
+		})
+	}
+}
+
+func BenchmarkQueue(b *testing.B) {
+	m := &mockMessage{
+		message: "foo",
+	}
+	q := NewPool(5)
+	q.Release()
+	for n := 0; n < b.N; n++ {
+		_ = q.Queue(m)
+	}
 }