## Before

[49894bf](https://github.com/golang-queue/queue/pull/72/commits/49894bf9a85a767e0564f8f8cc1ce75bb2bf60f0)

```sh
goos: linux
goarch: amd64
pkg: github.com/golang-queue/queue
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkQueueTask
INFO: 2022/05/29 04:18:27 logger.go:35: shutdown all tasks: 5 workers
BenchmarkQueueTask-2   	     584	   2079968 ns/op	    1801 B/op	      30 allocs/op
BenchmarkQueue
INFO: 2022/05/29 04:18:28 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 04:18:28 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 04:18:29 logger.go:35: shutdown all tasks: 5 workers
BenchmarkQueue-2       	 1892185	       624.9 ns/op	     214 B/op	       5 allocs/op
PASS
ok  	github.com/golang-queue/queue	2.924s
```

[5d2819e](https://github.com/golang-queue/queue/pull/72/commits/5d2819ea5e4b77e8dd7d8c915e9d2ab4c8370deb)

```sh
goos: linux
goarch: amd64
pkg: github.com/golang-queue/queue
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkQueueTask
BenchmarkQueueTask-2   	     582	   2055769 ns/op	    1403 B/op	      22 allocs/op
BenchmarkQueue
INFO: 2022/05/29 06:47:35 logger.go:35: shutdown all tasks: 4 workers
INFO: 2022/05/29 06:47:36 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 06:47:37 logger.go:35: shutdown all tasks: 4 workers
BenchmarkQueue-2       	 1996585	       616.1 ns/op	     212 B/op	       5 allocs/op
PASS
ok  	github.com/golang-queue/queue	2.932s
```

[15c43ec](https://github.com/golang-queue/queue/pull/72/commits/15c43eca41b777539a59eff02f47a1274c6d02b3) remove wait time in benchmark

## After

```sh
goos: linux
goarch: amd64
pkg: github.com/golang-queue/queue
cpu: Intel(R) Xeon(R) Platinum 8370C CPU @ 2.80GHz
BenchmarkQueueTask
INFO: 2022/05/29 14:53:36 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 14:53:36 logger.go:35: shutdown all tasks: 4 workers
INFO: 2022/05/29 14:53:37 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 14:53:38 logger.go:35: shutdown all tasks: 5 workers
BenchmarkQueueTask-2   	14850574	        86.25 ns/op	      54 B/op	       1 allocs/op
BenchmarkQueue
INFO: 2022/05/29 14:53:38 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 14:53:38 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 14:53:39 logger.go:35: shutdown all tasks: 5 workers
INFO: 2022/05/29 14:53:40 logger.go:35: shutdown all tasks: 5 workers
BenchmarkQueue-2       	 2270523	       484.3 ns/op	     191 B/op	       4 allocs/op
PASS
ok  	github.com/golang-queue/queue	3.738s
```